Renamed methods. Moved EnumOf to internals.
