+++ src/projects/EnsureThat/Enforcers/EnumArg.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+using EnsureThat.Internals;
 using JetBrains.Annotations;
 
 namespace EnsureThat.Enforcers
@@ -18,7 +16,7 @@ namespace EnsureThat.Enforcers
         /// Flags example:
         /// 
         /// [Flags]
-        /// enum Abc{
+        /// enum Abc {
         ///   A = 1,
         ///   B = 2,
         ///   C = 4,
@@ -28,7 +26,7 @@ namespace EnsureThat.Enforcers
         /// Abc.A | Abc.B IsDefined=true (due to Abc.AB)
         /// Abc.A | Abc.C IsDefined=false (A and C are both valid, the composite is valid due to <see cref="FlagsAttribute"/> attribute, but the composite is not a named enum value
         /// </example>
-        public T IsStrictlyDefined<T>(T value, [InvokerParameterName] string paramName = null, OptsFn optsFn = null) where T : struct, Enum
+        public T IsDefined<T>(T value, [InvokerParameterName] string paramName = null, OptsFn optsFn = null) where T : struct, Enum
         {
             if (!Enum.IsDefined(typeof(T), value))
             {
@@ -46,7 +44,7 @@ namespace EnsureThat.Enforcers
         /// Confirms that the <paramref name="value"/> is defined in the enum <typeparamref name="T"/>.
         /// Supports <see cref="FlagsAttribute"/> attribute.
         /// </summary>
-        public T IsDefined<T>(T value, [InvokerParameterName] string paramName = null, OptsFn optsFn = null) where T : struct, Enum
+        public T IsDefinedExtended<T>(T value, [InvokerParameterName] string paramName = null, OptsFn optsFn = null) where T : struct, Enum
         {
             var isEnumDefined = EnumOf<T>.Contains(value);
 
@@ -61,46 +59,5 @@ namespace EnsureThat.Enforcers
 
             return value;
         }
-
-        private static class EnumOf<T> where T : struct, Enum
-        {
-            internal static readonly Type EnumType;
-            private static readonly bool _hasFlags;
-            private static readonly List<ulong> _values;
-
-            static EnumOf()
-            {
-                EnumType = typeof(T);
-
-                _hasFlags = EnumType.GetTypeInfo().GetCustomAttributes<FlagsAttribute>(false).Any();
-
-                var enumValues = Enum.GetValues(EnumType);
-                _values = new List<ulong>(enumValues.Length);
-                foreach (var v in enumValues) 
-                    _values.Add(Convert.ToUInt64(v));
-            }
-
-            internal static bool Contains(T value)
-            {
-                if (!_hasFlags)
-                    return Enum.IsDefined(EnumType, value);
-
-                var raw = Convert.ToUInt64(value);
-                if (raw == 0)
-                    return Enum.IsDefined(EnumType, value);
-
-                ulong sum = 0;
-
-                foreach (var val in _values)
-                {
-                    if ((raw & val) == val)
-                        sum |= val;
-                }
-
-                return sum == raw;
-            }
-        }
     }
-
-
 }
