+++ src/projects/EnsureThat/EnsureThatEnumExtensions.cs
@@ -12,7 +12,7 @@
         /// Flags example:
         /// 
         /// [Flags]
-        /// enum Abc{
+        /// enum Abc {
         ///   A = 1,
         ///   B = 2,
         ///   C = 4,
@@ -22,14 +22,14 @@
         /// Abc.A | Abc.B IsDefined=true (due to Abc.AB)
         /// Abc.A | Abc.C IsDefined=false (A and C are both valid, the composite is valid due to `<see cref="System.FlagsAttribute"/> attribute, but the composite is not a named enum value
         /// </example>
-        public static void IsStrictlyDefined<T>(this in Param<T> param) where T : struct, System.Enum
-            => Ensure.Enum.IsStrictlyDefined(param.Value, param.Name, param.OptsFn);
+        public static void IsDefined<T>(this in Param<T> param) where T : struct, System.Enum
+            => Ensure.Enum.IsDefined(param.Value, param.Name, param.OptsFn);
 
         /// <summary>
         /// Confirms that the <paramref name="param.Value"/> is defined in the enum <typeparamref name="T"/>.
         /// Supports <see cref="System.FlagsAttribute"/> attribute.
         /// </summary>
-        public static void IsDefined<T>(this in Param<T> param) where T : struct, System.Enum
-            => Ensure.Enum.IsDefined(param.Value, param.Name, param.OptsFn);
+        public static void IsDefinedExtended<T>(this in Param<T> param) where T : struct, System.Enum
+            => Ensure.Enum.IsDefinedExtended(param.Value, param.Name, param.OptsFn);
     }
 }
